fix: extract frozen trade values to a service
