+++ src/components/Trade/TradeConfirm/TradeConfirm.tsx
@@ -19,7 +19,7 @@ import type { Swapper } from '@shapeshiftoss/swapper'
 import { type TradeTxs } from '@shapeshiftoss/swapper'
 import { TxStatus } from '@shapeshiftoss/unchained-client'
 import { useCallback, useEffect, useMemo, useState } from 'react'
-import { useFormContext, useWatch } from 'react-hook-form'
+import { useFormContext } from 'react-hook-form'
 import { useTranslate } from 'react-polyglot'
 import { useSelector } from 'react-redux'
 import { type RouterProps } from 'react-router-dom'
@@ -29,9 +29,8 @@ import { HelperTooltip } from 'components/HelperTooltip/HelperTooltip'
 import { Row } from 'components/Row/Row'
 import { SlideTransition } from 'components/SlideTransition'
 import { RawText, Text } from 'components/Text'
-import type { getTradeAmountConstants } from 'components/Trade/hooks/useGetTradeAmounts'
-import { useGetTradeAmounts } from 'components/Trade/hooks/useGetTradeAmounts'
 import { getSwapperManager } from 'components/Trade/hooks/useSwapper/swapperManager'
+import { useFrozenTradeValues } from 'components/Trade/TradeConfirm/useFrozenTradeValues'
 import { WalletActions } from 'context/WalletProvider/actions'
 import { useErrorHandler } from 'hooks/useErrorToast/useErrorToast'
 import { useLocaleFormatter } from 'hooks/useLocaleFormatter/useLocaleFormatter'
@@ -62,7 +61,6 @@ export const TradeConfirm = ({ history }: RouterProps) => {
   const {
     handleSubmit,
     setValue,
-    control,
     formState: { isSubmitting },
   } = useFormContext<TS>()
   const translate = useTranslate()
@@ -70,96 +68,26 @@ export const TradeConfirm = ({ history }: RouterProps) => {
   const [swapper, setSwapper] = useState<Swapper<ChainId>>()
   const flags = useSelector(selectFeatureFlags)
 
-  const formTrade = useWatch({ control, name: 'trade' })
-  const formFees = useWatch({ control, name: 'fees' })
-  const formSellAssetFiatRate = useWatch({ control, name: 'sellAssetFiatRate' })
-  const formFeeAssetFiatRate = useWatch({ control, name: 'feeAssetFiatRate' })
-  const formSlippage = useWatch({ control, name: 'slippage' })
-  const formBuyAssetAccountId = useWatch({ control, name: 'buyAssetAccountId' })
-  const formSellAssetAccountId = useWatch({ control, name: 'sellAssetAccountId' })
-  const formBuyTradeAsset = useWatch({ control, name: 'buyTradeAsset' })
-
-  const [frozenTradeAmountConstants, setFrozenTradeAmountConstants] =
-    useState<ReturnType<typeof getTradeAmountConstants>>()
-  const [frozenTrade, setFrozenTrade] = useState<TS['trade']>()
-  const [frozenFees, setFrozenFees] = useState<TS['fees']>()
-  const [frozenSellAssetFiatRate, setFrozenSellAssetFiatRate] = useState<TS['sellAssetFiatRate']>()
-  const [frozenFeeAssetFiatRate, setFrozenFeeAssetFiatRate] = useState<TS['feeAssetFiatRate']>()
-  const [frozenSlippage, setFrozenSlippage] = useState<TS['slippage']>()
-  const [frozenBuyAssetAccountId, setFrozenBuyAssetAccountId] = useState<TS['buyAssetAccountId']>()
-  const [frozenSellAssetAccountId, setFrozenSellAssetAccountId] =
-    useState<TS['sellAssetAccountId']>()
-  const [frozenBuyTradeAsset, setFrozenBuyTradeAsset] = useState<TS['buyTradeAsset']>()
-
-  const tradeAmountConstants = useGetTradeAmounts()
   const {
     number: { toFiat },
   } = useLocaleFormatter()
-
-  useEffect(() => {
-    !frozenTradeAmountConstants && setFrozenTradeAmountConstants(tradeAmountConstants)
-    !frozenTrade && setFrozenTrade(formTrade)
-    !frozenFees && setFrozenFees(formFees)
-    !frozenSellAssetFiatRate && setFrozenSellAssetFiatRate(formSellAssetFiatRate)
-    !frozenFeeAssetFiatRate && setFrozenFeeAssetFiatRate(formFeeAssetFiatRate)
-    !frozenSlippage && setFrozenSlippage(formSlippage)
-    !frozenBuyAssetAccountId && setFrozenBuyAssetAccountId(formBuyAssetAccountId)
-    !frozenSellAssetAccountId && setFrozenSellAssetAccountId(formSellAssetAccountId)
-    !frozenBuyTradeAsset && setFrozenBuyTradeAsset(formBuyTradeAsset)
-  }, [
-    formBuyAssetAccountId,
-    formBuyTradeAsset,
-    formFeeAssetFiatRate,
-    formFees,
-    formSellAssetAccountId,
-    formSellAssetFiatRate,
-    formSlippage,
-    formTrade,
-    frozenBuyAssetAccountId,
-    frozenBuyTradeAsset,
-    frozenFeeAssetFiatRate,
-    frozenFees,
-    frozenSellAssetAccountId,
-    frozenSellAssetFiatRate,
-    frozenSlippage,
-    frozenTrade,
-    frozenTradeAmountConstants,
-    tradeAmountConstants,
-  ])
-
-  // If an executed value exists we want to ignore any subsequent updates and use the executed value
-  const tradeAmounts = useMemo(
-    () => frozenTradeAmountConstants ?? tradeAmountConstants,
-    [frozenTradeAmountConstants, tradeAmountConstants],
-  )
-  const trade = useMemo(() => frozenTrade ?? formTrade, [frozenTrade, formTrade])
-  const fees = useMemo(() => frozenFees ?? formFees, [frozenFees, formFees])
-  const sellAssetFiatRate = useMemo(
-    () => frozenSellAssetFiatRate ?? formSellAssetFiatRate,
-    [frozenSellAssetFiatRate, formSellAssetFiatRate],
-  )
-  const feeAssetFiatRate = useMemo(
-    () => frozenFeeAssetFiatRate ?? formFeeAssetFiatRate,
-    [frozenFeeAssetFiatRate, formFeeAssetFiatRate],
-  )
-  const slippage = useMemo(() => frozenSlippage ?? formSlippage, [frozenSlippage, formSlippage])
-  const buyAssetAccountId = useMemo(
-    () => frozenBuyAssetAccountId ?? formBuyAssetAccountId,
-    [frozenBuyAssetAccountId, formBuyAssetAccountId],
-  )
-  const sellAssetAccountId = useMemo(
-    () => frozenSellAssetAccountId ?? formSellAssetAccountId,
-    [frozenSellAssetAccountId, formSellAssetAccountId],
-  )
-  const buyTradeAsset = useMemo(
-    () => frozenBuyTradeAsset ?? formBuyTradeAsset,
-    [frozenBuyTradeAsset, formBuyTradeAsset],
-  )
 
   const {
     state: { isConnected, wallet },
     dispatch,
   } = useWallet()
+
+  const {
+    tradeAmounts,
+    trade,
+    fees,
+    sellAssetFiatRate,
+    feeAssetFiatRate,
+    slippage,
+    buyAssetAccountId,
+    sellAssetAccountId,
+    buyTradeAsset,
+  } = useFrozenTradeValues()
 
   const defaultFeeAsset = useAppSelector(state =>
     selectFeeAssetByChainId(state, trade?.sellAsset?.chainId ?? ''),
