+++ src/components/Trade/hooks/useSwapper/utils.ts
@@ -126,6 +126,7 @@ export const getFormFees = ({
   feeAsset,
 }: GetFormFeesArgs): DisplayFeeData<KnownChainIds> => {
   const networkFeeCryptoHuman = fromBaseUnit(trade?.feeData?.networkFee, feeAsset.precision)
+  console.log('xxx fee asset calc', { feeAsset, trade, networkFeeCryptoHuman })
 
   const { chainNamespace } = fromAssetId(sellAsset.assetId)
   switch (chainNamespace) {
